feat: react-webapp-added

+++ src/main/java/com/employee/webapp/model/Employee.java
@@ -8,7 +8,7 @@ public class Employee {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id", nullable = false)
-    private Integer id;
+    private Long id;
 
     @Column(name = "first_name", nullable = false)
     private String firstName;
@@ -19,11 +19,11 @@ public class Employee {
     @Column(name = "email", nullable = false)
     private String email;
 
-    public Integer getId() {
+    public Long getId() {
         return this.id;
     }
 
-    public void setId(final Integer id) {
+    public void setId(final Long id) {
         this.id = id;
     }
